refactor(api): extract shared request guards (authenticate, observed, provideWorkspace)
The per-request enforcement helpers move to request-guards.ts so the
contract groups and the upcoming MCP protocol route compose one
authenticate path, one rate-limit path, and one wide-event envelope.
No behavior change.
